% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nisra_df.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(x, field)
}
\arguments{
\item{x}{A dataset created using \code{\link[=nisra_read_dataset]{nisra_read_dataset()}}}

\item{field}{The metadata field to read. See details for a list of fields.}
}
\value{
Metadata if the field is found, otherwise \code{NULL}
}
\description{
Retrieve metadata from a dataset, such as contact information, notes, and
official statistics status.
}
\details{
TODO: details on required fields
}
