% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nisra_df.R
\name{nisra_df}
\alias{nisra_df}
\title{Create a nisra_df object}
\usage{
nisra_df(..., meta = list())
}
\arguments{
\item{...}{named arguments to convert to columns, or an existing data-frame,
or a named list of columns}

\item{meta}{List of metadata fields}
}
\value{
A \code{nisra_df}, a \link[tibble:tbl_df-class]{tbl_df} with an additional
\code{"nisra_df"} class attribute
}
\description{
Create a nisra_df, a data-frame with additional metadata. This uses
\code{\link[vctrs:new_data_frame]{vctrs::new_data_frame()}} to convert the \code{...} argument to a data-frame.
}
