% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{nisra_get_boundaries}
\alias{nisra_get_boundaries}
\title{Attach boundaries to a dataset}
\usage{
nisra_get_boundaries(x)
}
\arguments{
\item{x}{A \code{nisra_df} dataset}
}
\value{
An \code{sf} object with all the columns of \code{x} and an additional geometry
column.
}
\description{
Attach boundaries to a dataset
}
