% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nisra_data_portal.R
\name{nisra_read_dataset}
\alias{nisra_read_dataset}
\title{Read NISRA data portal dataset}
\usage{
nisra_read_dataset(dataset_code, flush_cache = FALSE)
}
\arguments{
\item{dataset_code}{Dataset code}

\item{flush_cache}{Ignore cached values}
}
\value{
A tibble with the requested dataset. If \code{dataset_code}
is not found, an error will be thrown.
}
\description{
Fetch a dataset from the NISRA data portal using the dataset code. You
can search for a dataset using \code{\link[=nisra_search]{nisra_search()}}.
}
\examples{
claimant_count_lgd <- nisra_read_dataset("CCMLGD")
}
