% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nisrarr-package.R
\docType{package}
\name{nisrarr-package}
\alias{nisrarr}
\alias{nisrarr-package}
\title{nisrarr: Download Data from the NISRA Data Portal}
\description{
Download data from the Northern Ireland Statistics and Research Agency (NISRA) data portal, accessed at \url{https://data.nisra.gov.uk}. NISRA is a government agency and the principal source of official statistics and social research on Northern Ireland.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarkPaulin/nisrarr}
  \item \url{https://markpaulin.github.io/nisrarr/}
  \item Report bugs at \url{https://github.com/MarkPaulin/nisrarr/issues}
}

}
\author{
\strong{Maintainer}: Mark Paulin \email{paulin.mark@googlemail.com} [copyright holder]

}
\keyword{internal}
