% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nisra_data_portal.R
\name{nisra_read_dataset}
\alias{nisra_read_dataset}
\title{Read NISRA data portal dataset}
\usage{
nisra_read_dataset(dataset_code, flush_cache = FALSE)
}
\arguments{
\item{dataset_code}{Dataset code}

\item{flush_cache}{Ignore cached values}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the dataset.
}
\description{
Read NISRA data portal dataset
}
